feat: move pagination types to a separate file for better organization
