+++ src/features/attendance/types/attendance.ts
@@ -1,3 +1,5 @@
+import type { PaginationLinks, PaginationMeta } from '@/types/pagination';
+
 export type Tag = {
   tag_id: string;
   content: string;
@@ -15,28 +17,6 @@ export type Attendance = {
   deadline_date: string;
   expired: string;
   course: Course;
-};
-
-export type PaginationLinks = {
-  first: string;
-  last: string;
-  prev: string | null;
-  next: string | null;
-};
-
-export type PaginationMeta = {
-  current_page: number;
-  from: number;
-  last_page: number;
-  links: {
-    url: string | null;
-    label: string;
-    active: boolean;
-  }[];
-  path: string;
-  per_page: number;
-  to: number;
-  total: number;
 };
 
 export type AttendancesResponse = {
